renamed estimator classes for more clarity
